# Replace recordingSpan attributes implementation with slice of attributes