--- a/sdk/trace/trace_test.go
+++ b/sdk/trace/trace_test.go
@@ -1945,3 +1945,11 @@ func TestWithIDGenerator(t *testing.T) {
 		require.NoError(t, err)
 	}
 }
+
+func TestEmptyRecordingSpanAttributes(t *testing.T) {
+	assert.Nil(t, (&recordingSpan{}).Attributes())
+}
+
+func TestEmptyRecordingSpanDroppedAttributes(t *testing.T) {
+	assert.Equal(t, 0, (&recordingSpan{}).DroppedAttributes())
+}