… sdk/trace/attributesmap.go
… sdk/trace/attributesmap_test.go