This is an alternate implementation to https://github.com/open-telemetry/opentelemetry-go/pull/2555. It is preferred to that PR because it realizes similar gains but maintains the order of attributes (in the order the user added them).

Instead of an LRU strategy for cap-ing span attributes, comply with the specification and drop last added. To do this, the attributesmap is replaced with a slice of attributes.

- Resolves #2554
- Is the last item to address and closes #2402

### Testing performance comparison

This looks to have considerable memory and allocation reduction and faster computations when compared to [`main`](https://github.com/open-telemetry/opentelemetry-go/tree/d5292e3cd1c2ff1d759aacc5959243520000ce0a).

```
$ benchstat bench-main-53ead308 bench-recordingSpan-attrs-slice-326b6462

name                                          old time/op    new time/op    delta
SpanProcessor-8                                 14.9µs ±10%    11.3µs ±11%  -24.29%  (p=0.000 n=8+10)
SpanProcessorVerboseLogging-8                   15.5µs ± 3%    12.1µs ± 6%  -22.01%  (p=0.000 n=8+9)
StartEndSpan/AlwaysSample-8                      978ns ±14%     636ns ± 5%  -35.00%  (p=0.000 n=10+10)
StartEndSpan/NeverSample-8                       367ns ± 2%     313ns ± 2%  -14.96%  (p=0.000 n=8+9)
SpanSetAttributesOverCapacity-8                 23.1µs ± 5%     2.2µs ± 3%  -90.29%  (p=0.000 n=10+9)
SpanWithAttributes_4/AlwaysSample-8             1.85µs ± 5%    1.18µs ± 3%  -36.22%  (p=0.000 n=8+9)
SpanWithAttributes_4/NeverSample-8               513ns ± 7%     468ns ± 1%   -8.94%  (p=0.000 n=9+7)
SpanWithAttributes_8/AlwaysSample-8             2.68µs ± 5%    1.62µs ± 4%  -39.66%  (p=0.000 n=8+9)
SpanWithAttributes_8/NeverSample-8               679ns ± 5%     620ns ± 3%   -8.65%  (p=0.000 n=8+9)
SpanWithAttributes_all/AlwaysSample-8           2.09µs ± 7%    1.48µs ± 3%  -29.13%  (p=0.000 n=9+9)
SpanWithAttributes_all/NeverSample-8             566ns ± 5%     507ns ± 6%  -10.57%  (p=0.000 n=8+10)
SpanWithAttributes_all_2x/AlwaysSample-8        3.72µs ±12%    2.11µs ± 4%  -43.20%  (p=0.000 n=10+10)
SpanWithAttributes_all_2x/NeverSample-8          753ns ± 6%     692ns ± 6%   -8.15%  (p=0.000 n=8+10)
SpanWithEvents_4/AlwaysSample-8                 2.04µs ± 7%    1.50µs ± 2%  -26.61%  (p=0.000 n=10+7)
SpanWithEvents_4/NeverSample-8                   386ns ± 3%     322ns ± 3%  -16.46%  (p=0.000 n=9+10)
SpanWithEvents_8/AlwaysSample-8                 3.02µs ± 3%    2.27µs ± 1%  -24.75%  (p=0.000 n=9+8)
SpanWithEvents_8/NeverSample-8                   395ns ± 8%     333ns ± 3%  -15.68%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/AlwaysSample-8    1.26µs ± 8%    0.91µs ± 2%  -28.15%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/NeverSample-8      423ns ±10%     363ns ± 8%  -14.03%  (p=0.000 n=9+10)
SpanWithEvents_WithTimestamp/AlwaysSample-8     1.24µs ± 5%    0.92µs ± 2%  -25.81%  (p=0.000 n=10+8)
SpanWithEvents_WithTimestamp/NeverSample-8       470ns ± 2%     408ns ± 1%  -13.17%  (p=0.000 n=10+9)
TraceID_DotString-8                             82.2ns ± 5%    81.9ns ± 2%     ~     (p=0.897 n=10+8)
SpanID_DotString-8                              62.3ns ± 2%    61.5ns ± 2%     ~     (p=0.052 n=10+10)

name                                          old alloc/op   new alloc/op   delta
SpanProcessor-8                                 12.3kB ± 0%     8.4kB ± 0%  -31.29%  (p=0.000 n=10+10)
SpanProcessorVerboseLogging-8                   13.4kB ± 0%     9.6kB ± 0%  -28.62%  (p=0.000 n=10+9)
StartEndSpan/AlwaysSample-8                       816B ± 0%      432B ± 0%  -47.06%  (p=0.000 n=10+10)
StartEndSpan/NeverSample-8                        224B ± 0%      128B ± 0%  -42.86%  (p=0.000 n=10+10)
SpanSetAttributesOverCapacity-8                 15.4kB ± 0%     0.5kB ± 0%  -96.77%  (p=0.000 n=10+10)
SpanWithAttributes_4/AlwaysSample-8             1.73kB ± 0%    1.14kB ± 0%  -34.26%  (p=0.000 n=10+10)
SpanWithAttributes_4/NeverSample-8                480B ± 0%      384B ± 0%  -20.00%  (p=0.000 n=10+10)
SpanWithAttributes_8/AlwaysSample-8             2.43kB ± 0%    1.90kB ± 0%  -21.71%  (p=0.000 n=10+10)
SpanWithAttributes_8/NeverSample-8                736B ± 0%      640B ± 0%  -13.04%  (p=0.000 n=10+10)
SpanWithAttributes_all/AlwaysSample-8           1.90kB ± 0%    1.71kB ± 0%  -10.08%  (p=0.000 n=10+10)
SpanWithAttributes_all/NeverSample-8              544B ± 0%      448B ± 0%  -17.65%  (p=0.000 n=10+10)
SpanWithAttributes_all_2x/AlwaysSample-8        3.20kB ± 0%    3.06kB ± 0%   -4.62%  (p=0.000 n=10+10)
SpanWithAttributes_all_2x/NeverSample-8           864B ± 0%      768B ± 0%  -11.11%  (p=0.000 n=10+10)
SpanWithEvents_4/AlwaysSample-8                 1.50kB ± 0%    0.86kB ± 0%  -42.55%  (p=0.000 n=10+10)
SpanWithEvents_4/NeverSample-8                    224B ± 0%      128B ± 0%  -42.86%  (p=0.000 n=10+10)
SpanWithEvents_8/AlwaysSample-8                 2.21kB ± 0%    1.31kB ± 0%  -40.58%  (p=0.000 n=10+10)
SpanWithEvents_8/NeverSample-8                    224B ± 0%      128B ± 0%  -42.86%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/AlwaysSample-8      992B ± 0%      544B ± 0%  -45.16%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/NeverSample-8       240B ± 0%      144B ± 0%  -40.00%  (p=0.000 n=10+10)
SpanWithEvents_WithTimestamp/AlwaysSample-8     1.02kB ± 0%    0.57kB ± 0%  -44.09%  (p=0.000 n=10+10)
SpanWithEvents_WithTimestamp/NeverSample-8        264B ± 0%      168B ± 0%  -36.36%  (p=0.000 n=10+10)

name                                          old allocs/op  new allocs/op  delta
SpanProcessor-8                                   85.0 ± 0%      35.0 ± 0%  -58.82%  (p=0.000 n=10+10)
SpanProcessorVerboseLogging-8                     91.0 ± 0%      41.0 ± 0%  -54.95%  (p=0.000 n=10+10)
StartEndSpan/AlwaysSample-8                       7.00 ± 0%      2.00 ± 0%  -71.43%  (p=0.000 n=10+10)
StartEndSpan/NeverSample-8                        3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
SpanSetAttributesOverCapacity-8                    264 ± 0%         3 ± 0%  -98.86%  (p=0.000 n=10+10)
SpanWithAttributes_4/AlwaysSample-8               17.0 ± 0%       6.0 ± 0%  -64.71%  (p=0.000 n=10+10)
SpanWithAttributes_4/NeverSample-8                4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+10)
SpanWithAttributes_8/AlwaysSample-8               25.0 ± 0%       7.0 ± 0%  -72.00%  (p=0.000 n=10+10)
SpanWithAttributes_8/NeverSample-8                4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+10)
SpanWithAttributes_all/AlwaysSample-8             19.0 ± 0%       7.0 ± 0%  -63.16%  (p=0.000 n=10+10)
SpanWithAttributes_all/NeverSample-8              4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+10)
SpanWithAttributes_all_2x/AlwaysSample-8          30.0 ± 0%       8.0 ± 0%  -73.33%  (p=0.000 n=10+10)
SpanWithAttributes_all_2x/NeverSample-8           4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+10)
SpanWithEvents_4/AlwaysSample-8                   18.0 ± 0%       9.0 ± 0%  -50.00%  (p=0.000 n=10+10)
SpanWithEvents_4/NeverSample-8                    3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
SpanWithEvents_8/AlwaysSample-8                   27.0 ± 0%      14.0 ± 0%  -48.15%  (p=0.000 n=10+10)
SpanWithEvents_8/NeverSample-8                    3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/AlwaysSample-8      11.0 ± 0%       5.0 ± 0%  -54.55%  (p=0.000 n=10+10)
SpanWithEvents_WithStackTrace/NeverSample-8       4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+10)
SpanWithEvents_WithTimestamp/AlwaysSample-8       12.0 ± 0%       6.0 ± 0%  -50.00%  (p=0.000 n=10+10)
SpanWithEvents_WithTimestamp/NeverSample-8        5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
```