@MrAlias 2022-02-03
Here is a benchmark showing the comparison between main, https://github.com/open-telemetry/opentelemetry-go/pull/2555, and this PR

<table class='benchstat '>
<tr class='configs'><th><th><a href=https://github.com/open-telemetry/opentelemetry-go/tree/d5292e3cd1c2ff1d759aacc5959243520000ce0a><code>main</code></a><th><a href=https://github.com/open-telemetry/opentelemetry-go/pull/2555>#2555</a><th><a href=https://github.com/open-telemetry/opentelemetry-go/pull/2576>this PR</a>


<tbody>
<tr><th><th colspan='3' class='metric'>time/op
<tr><td>AttributesMapToKeyValue-8<td>3.21µs ±12%<td><td>
<tr><td>SpanProcessor-8<td>14.9µs ±10%<td>12.6µs ± 9%<td>11.3µs ±11%
<tr><td>SpanProcessorVerboseLogging-8<td>15.5µs ± 3%<td>12.8µs ±16%<td>12.1µs ± 6%
<tr><td>StartEndSpan/AlwaysSample-8<td>978ns ±14%<td>708ns ±24%<td>636ns ± 5%
<tr><td>StartEndSpan/NeverSample-8<td>367ns ± 2%<td>312ns ± 9%<td>313ns ± 2%
<tr><td>SpanSetAttributesOverCapacity-8<td>23.1µs ± 5%<td>2.3µs ±13%<td>2.2µs ± 3%
<tr><td>SpanWithAttributes_4/AlwaysSample-8<td>1.85µs ± 5%<td>1.28µs ±11%<td>1.18µs ± 3%
<tr><td>SpanWithAttributes_4/NeverSample-8<td>513ns ± 7%<td>481ns ± 5%<td>468ns ± 1%
<tr><td>SpanWithAttributes_8/AlwaysSample-8<td>2.68µs ± 5%<td>1.60µs ± 5%<td>1.62µs ± 4%
<tr><td>SpanWithAttributes_8/NeverSample-8<td>679ns ± 5%<td>644ns ± 7%<td>620ns ± 3%
<tr><td>SpanWithAttributes_all/AlwaysSample-8<td>2.09µs ± 7%<td>1.36µs ± 3%<td>1.48µs ± 3%
<tr><td>SpanWithAttributes_all/NeverSample-8<td>566ns ± 5%<td>508ns ± 7%<td>507ns ± 6%
<tr><td>SpanWithAttributes_all_2x/AlwaysSample-8<td>3.72µs ±12%<td>2.57µs ± 9%<td>2.11µs ± 4%
<tr><td>SpanWithAttributes_all_2x/NeverSample-8<td>753ns ± 6%<td>690ns ± 8%<td>692ns ± 6%
<tr><td>SpanWithEvents_4/AlwaysSample-8<td>2.04µs ± 7%<td>1.55µs ± 9%<td>1.50µs ± 2%
<tr><td>SpanWithEvents_4/NeverSample-8<td>386ns ± 3%<td>324ns ± 7%<td>322ns ± 3%
<tr><td>SpanWithEvents_8/AlwaysSample-8<td>3.02µs ± 3%<td>2.32µs ± 8%<td>2.27µs ± 1%
<tr><td>SpanWithEvents_8/NeverSample-8<td>395ns ± 8%<td>339ns ± 7%<td>333ns ± 3%
<tr><td>SpanWithEvents_WithStackTrace/AlwaysSample-8<td>1.26µs ± 8%<td>1.01µs ±12%<td>0.91µs ± 2%
<tr><td>SpanWithEvents_WithStackTrace/NeverSample-8<td>423ns ±10%<td>378ns ±16%<td>363ns ± 8%
<tr><td>SpanWithEvents_WithTimestamp/AlwaysSample-8<td>1.24µs ± 5%<td>0.98µs ± 9%<td>0.92µs ± 2%
<tr><td>SpanWithEvents_WithTimestamp/NeverSample-8<td>470ns ± 2%<td>418ns ± 8%<td>408ns ± 1%
<tr><td>TraceID_DotString-8<td>82.2ns ± 5%<td>84.2ns ±10%<td>81.9ns ± 2%
<tr><td>SpanID_DotString-8<td>62.3ns ± 2%<td>62.7ns ± 6%<td>61.5ns ± 2%
<tr><td>&nbsp;
</tbody>

<tbody>
<tr><th><th colspan='3' class='metric'>alloc/op
<tr><td>AttributesMapToKeyValue-8<td><td><td>
<tr><td>SpanProcessor-8<td>12.3kB ± 0%<td>8.6kB ± 0%<td>8.4kB ± 0%
<tr><td>SpanProcessorVerboseLogging-8<td>13.4kB ± 0%<td>9.7kB ± 0%<td>9.6kB ± 0%
<tr><td>StartEndSpan/AlwaysSample-8<td>816B ± 0%<td>448B ± 0%<td>432B ± 0%
<tr><td>StartEndSpan/NeverSample-8<td>224B ± 0%<td>128B ± 0%<td>128B ± 0%
<tr><td>SpanSetAttributesOverCapacity-8<td>15.4kB ± 0%<td>1.0kB ± 0%<td>0.5kB ± 0%
<tr><td>SpanWithAttributes_4/AlwaysSample-8<td>1.73kB ± 0%<td>1.28kB ± 0%<td>1.14kB ± 0%
<tr><td>SpanWithAttributes_4/NeverSample-8<td>480B ± 0%<td>384B ± 0%<td>384B ± 0%
<tr><td>SpanWithAttributes_8/AlwaysSample-8<td>2.43kB ± 0%<td>1.54kB ± 0%<td>1.90kB ± 0%
<tr><td>SpanWithAttributes_8/NeverSample-8<td>736B ± 0%<td>640B ± 0%<td>640B ± 0%
<tr><td>SpanWithAttributes_all/AlwaysSample-8<td>1.90kB ± 0%<td>1.34kB ± 0%<td>1.71kB ± 0%
<tr><td>SpanWithAttributes_all/NeverSample-8<td>544B ± 0%<td>448B ± 0%<td>448B ± 0%
<tr><td>SpanWithAttributes_all_2x/AlwaysSample-8<td>3.20kB ± 0%<td>2.83kB ± 0%<td>3.06kB ± 0%
<tr><td>SpanWithAttributes_all_2x/NeverSample-8<td>864B ± 0%<td>768B ± 0%<td>768B ± 0%
<tr><td>SpanWithEvents_4/AlwaysSample-8<td>1.50kB ± 0%<td>0.88kB ± 0%<td>0.86kB ± 0%
<tr><td>SpanWithEvents_4/NeverSample-8<td>224B ± 0%<td>128B ± 0%<td>128B ± 0%
<tr><td>SpanWithEvents_8/AlwaysSample-8<td>2.21kB ± 0%<td>1.33kB ± 0%<td>1.31kB ± 0%
<tr><td>SpanWithEvents_8/NeverSample-8<td>224B ± 0%<td>128B ± 0%<td>128B ± 0%
<tr><td>SpanWithEvents_WithStackTrace/AlwaysSample-8<td>992B ± 0%<td>560B ± 0%<td>544B ± 0%
<tr><td>SpanWithEvents_WithStackTrace/NeverSample-8<td>240B ± 0%<td>144B ± 0%<td>144B ± 0%
<tr><td>SpanWithEvents_WithTimestamp/AlwaysSample-8<td>1.02kB ± 0%<td>0.58kB ± 0%<td>0.57kB ± 0%
<tr><td>SpanWithEvents_WithTimestamp/NeverSample-8<td>264B ± 0%<td>168B ± 0%<td>168B ± 0%
<tr><td>TraceID_DotString-8<td><td><td>
<tr><td>SpanID_DotString-8<td><td><td>
<tr><td>&nbsp;
</tbody>

<tbody>
<tr><th><th colspan='3' class='metric'>allocs/op
<tr><td>AttributesMapToKeyValue-8<td><td><td>
<tr><td>SpanProcessor-8<td>85.0 ± 0%<td>45.0 ± 0%<td>35.0 ± 0%
<tr><td>SpanProcessorVerboseLogging-8<td>91.0 ± 0%<td>51.0 ± 0%<td>41.0 ± 0%
<tr><td>StartEndSpan/AlwaysSample-8<td>7.00 ± 0%<td>3.00 ± 0%<td>2.00 ± 0%
<tr><td>StartEndSpan/NeverSample-8<td>3.00 ± 0%<td>2.00 ± 0%<td>2.00 ± 0%
<tr><td>SpanSetAttributesOverCapacity-8<td>264 ± 0%<td>4 ± 0%<td>3 ± 0%
<tr><td>SpanWithAttributes_4/AlwaysSample-8<td>17.0 ± 0%<td>5.0 ± 0%<td>6.0 ± 0%
<tr><td>SpanWithAttributes_4/NeverSample-8<td>4.00 ± 0%<td>3.00 ± 0%<td>3.00 ± 0%
<tr><td>SpanWithAttributes_8/AlwaysSample-8<td>25.0 ± 0%<td>5.0 ± 0%<td>7.0 ± 0%
<tr><td>SpanWithAttributes_8/NeverSample-8<td>4.00 ± 0%<td>3.00 ± 0%<td>3.00 ± 0%
<tr><td>SpanWithAttributes_all/AlwaysSample-8<td>19.0 ± 0%<td>5.0 ± 0%<td>7.0 ± 0%
<tr><td>SpanWithAttributes_all/NeverSample-8<td>4.00 ± 0%<td>3.00 ± 0%<td>3.00 ± 0%
<tr><td>SpanWithAttributes_all_2x/AlwaysSample-8<td>30.0 ± 0%<td>6.0 ± 0%<td>8.0 ± 0%
<tr><td>SpanWithAttributes_all_2x/NeverSample-8<td>4.00 ± 0%<td>3.00 ± 0%<td>3.00 ± 0%
<tr><td>SpanWithEvents_4/AlwaysSample-8<td>18.0 ± 0%<td>10.0 ± 0%<td>9.0 ± 0%
<tr><td>SpanWithEvents_4/NeverSample-8<td>3.00 ± 0%<td>2.00 ± 0%<td>2.00 ± 0%
<tr><td>SpanWithEvents_8/AlwaysSample-8<td>27.0 ± 0%<td>15.0 ± 0%<td>14.0 ± 0%
<tr><td>SpanWithEvents_8/NeverSample-8<td>3.00 ± 0%<td>2.00 ± 0%<td>2.00 ± 0%
<tr><td>SpanWithEvents_WithStackTrace/AlwaysSample-8<td>11.0 ± 0%<td>6.0 ± 0%<td>5.0 ± 0%
<tr><td>SpanWithEvents_WithStackTrace/NeverSample-8<td>4.00 ± 0%<td>3.00 ± 0%<td>3.00 ± 0%
<tr><td>SpanWithEvents_WithTimestamp/AlwaysSample-8<td>12.0 ± 0%<td>7.0 ± 0%<td>6.0 ± 0%
<tr><td>SpanWithEvents_WithTimestamp/NeverSample-8<td>5.00 ± 0%<td>4.00 ± 0%<td>4.00 ± 0%
<tr><td>TraceID_DotString-8<td><td><td>
<tr><td>SpanID_DotString-8<td><td><td>
<tr><td>&nbsp;
</tbody>

</table>